pyproject.toml
[tool.poetry]
name = "bas-python-api"
version = "0.1.0"
description = "Python library, which allows you to automate Google Chrome browser via BAS software."
authors = ["Sergerdn <example@example.com>"]
readme = "README.md"
packages = [
    { include = "bas_api" }
]

[tool.poetry.dependencies]
python = "^3.7"
bas-remote-python-v2 = { git = "https://github.com/sergerdn/bas-remote-python-v2" }
psutil = "^5.9.2"
pywin32 = "^304"
pydantic = "^1.10.2"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
twine = "^4.0.1"
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"
pytest-dependency = "^0.5.1"
setuptools = "^65.3.0"
python-dotenv = "^0.21.0"
tox = "^3.26.0"
tox-poetry = "^0.4.1"
flake8-tidy-imports = "^4.8.0"
flake8 = "^5.0.4"
mypy = "^0.971"
isort = "^5.10.1"
black = { git = "https://github.com/psf/black" }
poetryup = "^0.11.0"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'

[tool.poetry.build]
#script = "build.py"
generate-setup-file = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

